api_rest/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
pydantic
bcrypt
yfinance
PyJWT

